move adapter to seperate directory
