+++ daiquiri_core/adapter/mysql.py
@@ -1,40 +1,9 @@
 import json
 import re
 
-from django.conf import settings
 from django.db import connections
 
-
-def get_adapter(database_config):
-    if database_config in settings.DATABASES:
-        if settings.DATABASES[database_config]['ENGINE'] == 'django.db.backends.mysql':
-            return MySQLAdapter(database_config)
-        else:
-            raise Exception('database engine "%s" is not supported (yet)' % settings.DATABASES[database_config]['ENGINE'])
-    else:
-        raise Exception('no database config named "%s" found in settings.DATABASES' % database_config)
-
-
-class BaseAdapter(object):
-
-    def execute(self, sql):
-        return self.cursor.execute(sql)
-
-    def fetchone(self, sql, args=None):
-        if args:
-            self.cursor.execute(sql, args)
-        else:
-            self.cursor.execute(sql)
-
-        return self.cursor.fetchone()
-
-    def fetchall(self, sql, args=None):
-        if args:
-            self.cursor.execute(sql, args)
-        else:
-            self.cursor.execute(sql)
-
-        return self.cursor.fetchall()
+from .base import BaseAdapter
 
 
 class MySQLAdapter(BaseAdapter):
